Rename minerva2 package to minerva
